refactor: add source module entry
